requirements.txt
python-dotenv

pygame
pyglet

pywebio

# rainbowio
adafruit-circuitpython-seesaw
